+++ Example/ItemSetServer/Tasks.h
@@ -27,12 +27,15 @@ struct TaskRecord : Ramio::MetaItemData
 	RMetaInt state = 0;
 	RMetaString color;
 
+	QVariant stateStr(const TaskRecord& data) const;
+
 	RMETA_OBJECT_START(MetaItemData)
 	RMETA_OBJECT_FIELD(text, String, "Text", "Текст", Field)
 	RMETA_OBJECT_FIELD(time, DateTime, "Time", "Время", Field)
 	RMETA_OBJECT_FIELD(userId, PKey, "UserId", "Пользователь", FKey)
 	RMETA_OBJECT_FIELD(state, Int, "State", "Состояние", Field)
 	RMETA_OBJECT_FIELD(color, String, "Color", "Цвет", Field)
+	RMETA_OBJECT_FUNCTION(TaskRecord, stateStr, String, "StateStr", "Состояние", Function)
 	RMETA_OBJECT_END
 
 	TaskRecord() = default;
@@ -55,4 +58,6 @@ public:
 };
 
 
-GENERATE_HEADER_CLASS_METASET(MetaTaskSet, Task, TaskRecord)
+GENERATE_HEADER_CLASS_METASET_START(MetaTaskSet, Task, TaskRecord)
+std::function<QString(const Ramio::BaseMetaItemData&)> colorFunction;
+};